split pong file
